+++ CHANGELOG.md
@@ -6,6 +6,7 @@ TBD
 
 - Replication of all events, when configured, except for Garden Sync events. Garden sync events spawn off additional events that will be replicated properly.
 - Add support for commands to have tags, used for filtering on the commands index page
+- Publish Instance Status changed to Instance default topic {Namespace}.{System}.{Version}.{Instance}
 - Upgraded Brewtils version to [3.22.0](https://github.com/beer-garden/brewtils/releases/tag/3.22.0)
 
 # 3.21.0